AntescofoImport head set to the old Ircam Forge’s Head of OFx90 Branch. I had actually copied the Master Branch which was incorrect! Compiles perfectly!

+++ Repeat.cpp
@@ -17,7 +17,7 @@
 
 using namespace antescofo;
 
-Repeat::Repeat( int measure, int direction, int ending ):
+Repeat::Repeat( float measure, int direction, int ending ):
     Event  (),
     direction_ ( direction ),
     ending_    ( ending )
